Reduce memory usage, allow selection of multiple metrics in MetricsPane [HZG-208] (#7)
This PR reduces the memory footprint while loading large diagnostics
files.

Previously each individual `DiagnosticFile` contained an index for the
metric entries contained within it. Now there is a single index per
instance of `InstanceDiagnostics`, this hugely reduces the number of
index entries needed and allows the tool to load several gigabytes of
diagnostic logs.

In addition a number of IDE compiler warnings have been fixed and the
`MetricsPane` has been updated to allow selection and comparison of
multiple metrics.